[PM-2807] Refactor peergroup util spec and make it private
