+++ scalanet/src/io/iohk/scalanet/peergroup/NettyFutureUtils.scala
@@ -1,17 +1,12 @@
 package io.iohk.scalanet.peergroup
 
-import java.net.{InetSocketAddress, ServerSocket}
 import io.netty
 import io.netty.util.concurrent.{Future, GenericFutureListener}
 import monix.eval.Task
 
 import java.util.concurrent.CancellationException
 
-object InetPeerGroupUtils {
-
-  type ChannelId = (InetSocketAddress, InetSocketAddress)
-
-  // TODO: This has nothing to do with InetPeerGroup, move it somewhere else.
+private[scalanet] object NettyFutureUtils {
   def toTask(f: => netty.util.concurrent.Future[_]): Task[Unit] = {
     fromNettyFuture(Task.delay(f)).void
   }
@@ -40,19 +35,5 @@ object InetPeerGroupUtils {
       }
     })
     ()
-  }
-
-  def getChannelId(remoteAddress: InetSocketAddress, localAddress: InetSocketAddress): ChannelId = {
-    (remoteAddress, localAddress)
-  }
-
-  // TODO: Can we move this to tests?
-  def aRandomAddress(): InetSocketAddress = {
-    val s = new ServerSocket(0)
-    try {
-      new InetSocketAddress("localhost", s.getLocalPort)
-    } finally {
-      s.close()
-    }
   }
 }
